show outdated fusdc farm inline
